deletes working again

+++ kanban.server/server/services/BoardsService.js
@@ -2,8 +2,8 @@ import { dbContext } from '../db/DbContext'
 import { BadRequest } from '../utils/Errors'
 
 class BoardsService {
-  async delete(id) {
-    const data = await dbContext.Boards.findOneAndDelete(id)
+  async delete(id, userId) {
+    const data = await dbContext.Boards.findOneAndDelete({ _id: id, creatorId: userId })
     if (!data) {
       throw new BadRequest('Invalid Id')
     }
